Add randomization util and make randomization method configurable
