+++ sam.local.yml
@@ -20,3 +20,4 @@ Resources:
           S3_FEED_KEY: 'feed.xml'
           LOG_LEVEL: debug
           TZ: 'America/New_York'
+          RANDOMIZATION_METHOD: 'uniform'
